changed from long to int

+++ assignments/2-stdio/checker-lin/libso_stdio.c
@@ -171,7 +171,7 @@ long so_ftell(SO_FILE *stream)
 size_t so_fread(void *ptr, size_t size, size_t nmemb, SO_FILE *stream)
 {
     size_t cnt = 0;
-    long var = 0;
+    int var = 0;
 
     while (cnt < size * nmemb) {
         var = so_fgetc(stream);
@@ -190,7 +190,7 @@ size_t so_fread(void *ptr, size_t size, size_t nmemb, SO_FILE *stream)
 size_t so_fwrite(const void *ptr, size_t size, size_t nmemb, SO_FILE *stream)
 {
     size_t cnt = 0;
-    long var = 0;
+    int var = 0;
 
     while (cnt < size * nmemb) {
         unsigned char character = *(unsigned char *)(ptr + cnt);
